Update GTrends.php
BUGFIX: If we transform the JSON as an array, some entries inside the JSON provided by Google Trends will be transform from an object to an array. This is a problem when we are doing request from international trends.

The second request will get an 400 error.
To avoid the problem, we need to work the JSON content as object format (default format used by Json::decode).